+++ example/lib/screens/inputs.dart
@@ -1,5 +1,4 @@
 import 'package:fluent_ui/fluent_ui.dart';
-// import 'package:flutter/material.dart' as m;
 
 class InputsPage extends StatefulWidget {
   const InputsPage({Key? key}) : super(key: key);